langchain-nfe/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
langchain
langchain-google-genai
streamlit
langchain_experimental
tabulate
PyMuPDF>=1.26.5
openpyxl>=3.1.5
python_docx>=1.2.0